fps limiter, modelloader

+++ masterproj/vulkan-proj/renderer/vulkan/vInitializer.h
@@ -19,6 +19,7 @@
 #include <set>
 #include <fstream>
 #include <chrono>
+#include <thread>
 
 #include "../VideoInfo.h"
 #include "../../model/ModelLoader.h"
@@ -202,7 +203,7 @@ private:
 	std::vector<VkSemaphore> imageAvailableSemaphores;
 	std::vector<VkSemaphore> renderFinishedSemaphores;
 	std::vector<VkFence> inFlightFences;
-	const int MAX_FRAMES_IN_FLIGHT = 2;
+	const size_t MAX_FRAMES_IN_FLIGHT = 2;
 	size_t currentFrame = 0;
 
 	/* Swap chain recreation */
